raspberry pi/rpi_requirements.txt
opencv-contrib-python==4.5.3.56
picamera== 1.13
tflite-runtime == 2.9.1
Pillow == 8.1.2
numpy == 1.23.2